+++ src/ui/panes/queue.rs
@@ -90,7 +90,8 @@ impl QueuePane {
                 .theme
                 .song_table_format
                 .iter()
-                .map(|v| Into::<Constraint>::into(v.width))
+                // This 0 is fine - song_table_format should never have the Ratio constraint
+                .map(|v| v.width.into_constraint(0))
                 .collect_vec(),
             context.config.theme.song_table_format.iter().map(|v| v.prop.clone()).collect_vec(),
         )